Split longs table into separate components
